requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dill==0.4.0
fastdtw==0.3.4
filelock==3.18.0
Flask==2.3.2
flatbuffers==25.2.10
fsspec==2025.7.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.68.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
keras==3.11.2
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.1.0
narwhals==2.1.2
networkx==3.5
ntlm-auth==1.5.0
numpy==1.23.5
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==1.5.3
pbr==6.1.1
pillow==11.3.0
ply==3.11
protobuf==4.25.8
psutil==7.0.0
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
qiskit==0.43.1
qiskit-aer==0.12.0
qiskit-ibm-provider==0.6.1
qiskit-ibm-runtime==0.11.2
qiskit-ibmq-provider==0.20.2
qiskit-machine-learning==0.6.1
qiskit-terra==0.24.1
referencing==0.36.2
requests==2.32.4
requests-ntlm==1.1.0
rich==14.1.0
rpds-py==0.27.0
rustworkx==0.16.0
scikit-learn==1.2.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
stevedore==5.4.1
streamlit==1.48.1
symengine==0.9.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
